+++ achievements/kills/achievements_kill_elite.lua
@@ -1,11 +1,12 @@
 ------------------------------------------------------
 -- Named Kill Bosses / Elites
+-- NOTE: Descriptions changed to "witness / be present" wording
 ------------------------------------------------------
 
 table.insert(KAMN.achievements, {
   id = "KILL_HOGGER",
   name = "Justice for Elwynn",
-  description = "Defeat Hogger, the terror of Elwynn Forest.",
+  description = "Witness the defeat of Hogger, the terror of Elwynn Forest.",
   type = "namedkill",
   alttype = "elite",
   targetname = "Hogger",
@@ -20,7 +21,7 @@ table.insert(KAMN.achievements, {
 table.insert(KAMN.achievements, {
   id = "KILL_BELLYGRUB",
   name = "Boarbecue",
-  description = "Defeat Bellygrub, the infamous gluttonous boar.",
+  description = "Witness the defeat of Bellygrub, the infamous gluttonous boar.",
   type = "namedkill",
   alttype = "elite",
   targetname = "Bellygrub",
@@ -35,7 +36,7 @@ table.insert(KAMN.achievements, {
 table.insert(KAMN.achievements, {
   id = "KILL_VANCLEEF",
   name = "Brotherhood Broken",
-  description = "Defeat Edwin VanCleef, leader of the Defias Brotherhood.",
+  description = "Witness the defeat of Edwin VanCleef, leader of the Defias Brotherhood.",
   type = "namedkill",
   alttype = "elite",
   targetname = "Edwin VanCleef",
@@ -50,7 +51,7 @@ table.insert(KAMN.achievements, {
 table.insert(KAMN.achievements, {
   id = "BOSS_SHARPTUSK",
   name = "Sharptusk Falls",
-  description = "Defeat Chief Sharptusk Thornmantle.",
+  description = "Witness the defeat of Chief Sharptusk Thornmantle.",
   type = "namedkill",
   alttype = "elite",
   targetname = "Chief Sharptusk Thornmantle",
@@ -64,12 +65,13 @@ table.insert(KAMN.achievements, {
 
 ------------------------------------------------------
 -- Named Kill Bosses – Classic Raid Targets
+-- NOTE: Descriptions changed to "witness / be present" wording
 ------------------------------------------------------
 
 table.insert(KAMN.achievements, {
   id = "KILL_RAGNAROS",
   name = "The Firelord Falls",
-  description = "Defeat Ragnaros, the Firelord of Molten Core.",
+  description = "Be present for the defeat of Ragnaros, the Firelord of Molten Core.",
   type = "bosskill",
   alttype = "boss",
   targetname = "Ragnaros",
@@ -84,7 +86,7 @@ table.insert(KAMN.achievements, {
 table.insert(KAMN.achievements, {
   id = "KILL_NEFARIAN",
   name = "End of the Blackwing",
-  description = "Defeat Nefarian in Blackwing Lair.",
+  description = "Be present for the defeat of Nefarian in Blackwing Lair.",
   type = "bosskill",
   alttype = "boss",
   targetname = "Nefarian",
@@ -99,7 +101,7 @@ table.insert(KAMN.achievements, {
 table.insert(KAMN.achievements, {
   id = "KILL_CTHUN",
   name = "Eye for an Eye",
-  description = "Defeat C'Thun in the Temple of Ahn'Qiraj.",
+  description = "Be present for the defeat of C'Thun in the Temple of Ahn'Qiraj.",
   type = "bosskill",
   alttype = "boss",
   targetname = "C'Thun",
@@ -114,7 +116,7 @@ table.insert(KAMN.achievements, {
 table.insert(KAMN.achievements, {
   id = "KILL_KELTHUZAD",
   name = "The Lich's End",
-  description = "Defeat Kel'Thuzad in Naxxramas.",
+  description = "Be present for the defeat of Kel'Thuzad in Naxxramas.",
   type = "bosskill",
   alttype = "boss",
   targetname = "Kel'Thuzad",
@@ -129,7 +131,7 @@ table.insert(KAMN.achievements, {
 table.insert(KAMN.achievements, {
   id = "KILL_ONYXIA",
   name = "Onyxia's Head Delivered",
-  description = "Defeat Onyxia in her lair.",
+  description = "Be present for the defeat of Onyxia in her lair.",
   type = "bosskill",
   alttype = "boss",
   targetname = "Onyxia",
@@ -144,7 +146,7 @@ table.insert(KAMN.achievements, {
 table.insert(KAMN.achievements, {
   id = "KILL_HAKKAR",
   name = "The Soulflayer's Doom",
-  description = "Defeat Hakkar the Soulflayer in Zul'Gurub.",
+  description = "Be present for the defeat of Hakkar the Soulflayer in Zul'Gurub.",
   type = "bosskill",
   alttype = "boss",
   targetname = "Hakkar",
@@ -159,7 +161,7 @@ table.insert(KAMN.achievements, {
 table.insert(KAMN.achievements, {
   id = "KILL_MANDOKIR",
   name = "Heads Will Roll",
-  description = "Defeat Bloodlord Mandokir in Zul'Gurub.",
+  description = "Be present for the defeat of Bloodlord Mandokir in Zul'Gurub.",
   type = "bosskill",
   alttype = "boss",
   targetname = "Bloodlord Mandokir",
@@ -174,7 +176,7 @@ table.insert(KAMN.achievements, {
 table.insert(KAMN.achievements, {
   id = "KILL_JINDO",
   name = "Hex Breaker",
-  description = "Defeat Jin'do the Hexxer in Zul'Gurub.",
+  description = "Be present for the defeat of Jin'do the Hexxer in Zul'Gurub.",
   type = "bosskill",
   alttype = "boss",
   targetname = "Jin'do the Hexxer",
@@ -189,7 +191,7 @@ table.insert(KAMN.achievements, {
 table.insert(KAMN.achievements, {
   id = "KILL_AZUREGOS",
   name = "Blue Dragon Down",
-  description = "Defeat the world boss Azuregos in Azshara.",
+  description = "Be present for the defeat of the world boss Azuregos in Azshara.",
   type = "bosskill",
   alttype = "boss",
   targetname = "Azuregos",
@@ -204,7 +206,7 @@ table.insert(KAMN.achievements, {
 table.insert(KAMN.achievements, {
   id = "KILL_KAZZAK",
   name = "Lord of Doom",
-  description = "Defeat Lord Kazzak in the Blasted Lands.",
+  description = "Be present for the defeat of Lord Kazzak in the Blasted Lands.",
   type = "bosskill",
   alttype = "boss",
   targetname = "Lord Kazzak",